refactor: ♻️ convert to using updated template data package (#29)
# Description

This PR updates the repo to use the updated `template-data-package`. It
also starts the process of using the `*_properties.py` scripts in
`main.py`.

For this PR, I've followed the Sprout [Creating a Data
Package](https://sprout.seedcase-project.org/docs/guide/packages) guide
and updated the package properties and `datapackage.json` and
`README.md` by uncommenting those steps in `main.py`.

I'll do a separate PR that adds the data resources and updates the
preprocessing scripts.

Closes #23 

This PR needs an in-depth review.

## Checklist

- [X] Ran `just run-all` (fails bc some comment lines are too long in
`scripts/download-data.py` and `scripts/convert-meta.py`, will fix in a
separate PR adding the data resources)

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>